+++ contracts/contracts/l2/staking/Gov.sol
@@ -7,7 +7,6 @@ import {EnumerableSetUpgradeable} from "@openzeppelin/contracts-upgradeable/util
 import {Predeploys} from "../../libraries/constants/Predeploys.sol";
 import {ISequencer} from "./ISequencer.sol";
 import {IGov} from "./IGov.sol";
-import {IL2Staking} from "./IL2Staking.sol";
 
 contract Gov is IGov, OwnableUpgradeable {
     using EnumerableSetUpgradeable for EnumerableSetUpgradeable.AddressSet;
@@ -297,7 +296,7 @@ contract Gov is IGov, OwnableUpgradeable {
         address[] memory latestSequencerSet = ISequencer(SEQUENCER_CONTRACT)
             .getSequencerSet2();
         uint256 validVotes = 0;
-        for (uint i = 0; i < latestSequencerSet.length; i++) {
+        for (uint256 i = 0; i < latestSequencerSet.length; i++) {
             if (votes[proposalID].contains(latestSequencerSet[i])) {
                 validVotes = validVotes + 1;
             }
