+++ contracts/contracts/test/Gov.t.sol
@@ -2,7 +2,6 @@
 pragma solidity =0.8.24;
 
 import {L2StakingBaseTest} from "./base/L2StakingBase.t.sol";
-import {Gov} from "../l2/staking/Gov.sol";
 import {IGov} from "../l2/staking/IGov.sol";
 import {Types} from "../libraries/common/Types.sol";
 import {ICrossDomainMessenger} from "../libraries/ICrossDomainMessenger.sol";
@@ -183,7 +182,7 @@ contract GovTest is L2StakingBaseTest {
         assertFalse(canBeApproved);
 
         // invalide votes
-        for (uint i = 0; i < removed.length - 1; i++) {
+        for (uint256 i = 0; i < removed.length - 1; i++) {
             assertTrue(gov.isVoted(currentproposalID, removed[i]));
         }
